fix: theme 적용을 위한 decorator 추가
