+++ frontend/.storybook/preview.tsx
@@ -1,0 +1,26 @@
+import GlobalStyles from "../src/styles/globalStyles";
+import { theme } from "../src/styles/theme";
+import type { Preview } from "@storybook/react";
+import React from "react";
+import { ThemeProvider } from "styled-components";
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+};
+
+export const decorators = [
+  (Story) => (
+    <ThemeProvider theme={theme}>
+      <GlobalStyles />
+      <Story />
+    </ThemeProvider>
+  ),
+];
+export default preview;
